refactoring: moved definitions into another file
