+++ packages/typir/src/utils/utils.ts
@@ -69,3 +69,17 @@ export function assertTypirType<T extends Type>(type: unknown, check: (type: unk
         throw new Error(msg ?? `'${type}' has another type`);
     }
 }
+
+
+/**
+ * A deep partial type definition for services. We look into T to see whether its type definition contains
+ * any methods. If it does, it's one of our services and therefore should not be partialized.
+ * Copied from Langium.
+ */
+//eslint-disable-next-line @typescript-eslint/ban-types
+export type DeepPartial<T> = T[keyof T] extends Function ? T : {
+    [P in keyof T]?: DeepPartial<T[P]>;
+};
+
+/** Makes only the specified properties of the given type optional */
+export type MakePropertyOptional<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
